fix(build): resolve Kotlin compilation errors in MedicalDatabaseTools, KoogAgentFactory, OllamaProvider, and ChatViewModel

+++ app/src/main/java/com/clinref/app/domain/ai/providers/OllamaProvider.kt
@@ -32,7 +32,7 @@ class OllamaProvider(
                 ai.koog.prompt.llm.LLMCapability.Tools,
                 ai.koog.prompt.llm.LLMCapability.Temperature
             ),
-            contextLength = contextLength
+            contextLength = contextLength.toLong()
         )
     }
